+++ src/test/java/TestStartup.java
@@ -1,5 +1,6 @@
 import com.heybcat.docker.pull.CommandStartup;
 import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
 import java.net.URISyntaxException;
 import java.util.concurrent.locks.LockSupport;
 import org.junit.Test;
@@ -11,11 +12,8 @@ public class TestStartup {
         try {
             CommandStartup.main(new String[]{"--web"});
             LockSupport.park();
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        } catch (InterruptedException e) {
+        } catch (URISyntaxException | InvocationTargetException | NoSuchMethodException | IllegalAccessException |
+                 InterruptedException | IOException e) {
             throw new RuntimeException(e);
         }
     }
